==> 0_sel_zomato_url.ipynb <==
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import TimeoutException
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.chrome.options import Options
import time
import calendar
import pandas  as pd 
from tqdm import tqdm

1. extract restaurant url from each region in jakarta.
2. save as txt 

def extract_main_page(url):
#     url = 'https://www.zomato.com/jakarta/kelapa-gading-restaurants'
    driver = webdriver.Chrome(executable_path='chromedriver.exe')

    chrome_options = Options()
    chrome_options.add_argument("--disable-popup-blocking")
    driver.get(url)
    
    time.sleep(3)
    # scroll until the end of the page
    
    # method 1 
#     page = driver.find_element_by_tag_name("html")
#     page.send_keys(Keys.END)

#     time.sleep(120)
    
    # method 2 
    
    last_height = 0
    while True:
        # Scroll down to bottom
        driver.execute_script("window.scrollBy({}, {});".format(last_height,last_height+100))
        # Wait to load page
        time.sleep(5)
        # Calculate new scroll height and compare with last scroll height
        new_height = driver.execute_script("return document.body.scrollHeight")
#         print(last_height, new_height,'ok')
        if new_height == last_height:
            break
        last_height = new_height
    
    # get all restaurants url 
    lnks=driver.find_elements_by_xpath("//a[@href]")
    restaurant_link = [] 

    for lnk in lnks:
        if 'info' in lnk.get_attribute("href"): 
            restaurant_link.append(lnk.get_attribute("href"))
        else : 
            pass 
    restaurant_link = list(set(restaurant_link))
    
    driver.close()
    return restaurant_link

df = pd.read_csv('data/0_jak_area.txt')
df = df.tail(6)

df.tail(6)

df['url'].loc[0]

restaurant_list = []
for item in tqdm(df.itertuples()):
    tmp = {} 
    tmp['url'] = item.url
    tmp['link'] = extract_main_page(item.url)
    restaurant_list.append(tmp)
    time.sleep(3)

# restaurant_list[0]

for item in restaurant_list:
    print(item.keys(), len(list(item.values())[0]))

# for key, value in restaurant_list.items():
#     print (key, len(value))

tes = pd.DataFrame(restaurant_list)
tes['total'] = tes['link'].apply(len)

tes['total'].sum()

tes

tes.iloc[2]['url']

tes.to_pickle('data/restaurant_raw2_extra.pickle')

tes.to_csv('data/1_bodebek.csv')
