==> 3_geo_extract.ipynb <==
import pandas as pd 
from geolib import geohash
from haversine import haversine, Unit
from math import radians, cos, sin, asin, sqrt
from scipy.stats import entropy
import numpy as np 
from numpy.linalg import norm

def calculate_distance(lat1, lon1, lat2, lon2 ):
    # haversine formula 


    lon1, lat1, lon2, lat2 = map(np.radians, [lon1, lat1, lon2, lat2])

    dlon = lon2 - lon1
    dlat = lat2 - lat1

    a = np.sin(dlat/2.0)**2 + np.cos(lat1) * np.cos(lat2) * np.sin(dlon/2.0)**2

    c = 2 * np.arcsin(np.sqrt(a))
    km = 6367 * c * 1000 # in meter 
    return km

def compute_entropy(df_input):
    
    return entropy(df_input)


def density_gof(df_zomato, df_poi, distance):
    # data preparation 
    df_zom_geo = df_zomato[['index','lat','long','geohash']]
    df_zom_geo['neighbors'] = df_zom_geo['geohash'].apply(lambda x:list(geohash.neighbours(x)))
    df_zom_geo = df_zom_geo.explode('neighbors')

    # merge with poi 
    df_zom_poi = pd.merge(df_zom_geo, df_poi, left_on = 'neighbors', right_on = 'geohash6', how='left')
    df_zom_poi['distance'] = df_zom_poi.apply(lambda x:calculate_distance(x['lat'],x['long'],x['lat_poi'],x['long_poi']), axis= 1)

    # filter distance 
    df_zom_poi = df_zom_poi[df_zom_poi['distance'] < distance]

    # aggregate  
    df_zom_poi_gb = df_zom_poi.groupby(['index_x','poi_type']).agg({'index_y':'nunique'}).reset_index()
    df_zom_poi_gb.columns = ['index','poi_type','total']
    df_zom_poi_gb['total'] = df_zom_poi_gb['total'].astype(int)

    # pivot 
    df_zom_poi_pivot = df_zom_poi_gb.pivot_table(index='index', columns='poi_type', values='total').fillna(0).reset_index()
    prefix_name = 'd_{}_'.format(distance)
    df_zom_poi_pivot = df_zom_poi_pivot.add_prefix(prefix_name)

    return df_zom_poi_pivot 

def compt_gof(df_zomato_1, df_zomato_2, distance):
    # data preparation 
    df_zom_geo = df_zomato_1[['index','lat','long','geohash','encode','rating']]
    df_zom_geo['neighbors'] = df_zom_geo['geohash'].apply(lambda x:list(geohash.neighbours(x)))
    df_zom_geo = df_zom_geo.explode('neighbors')

    #Preprocessing for df_zomato_2

    df_zom_geo2 = df_zomato_2[['index','lat','long','geohash','encode','rating']]
    df_zom_geo2.columns = ['index2','lat2','long2','geohash2','encode2','rating2']


    # self merging
    df_zom_zom = pd.merge(df_zom_geo, df_zom_geo2, left_on = 'neighbors', right_on = 'geohash2', how='inner')
    df_zom_zom['distance'] = df_zom_zom.apply(lambda x:calculate_distance(x['lat'],x['long'],x['lat2'],x['long2']), axis= 1)

    # filter distance 
    df_zom_zom = df_zom_zom[df_zom_zom['distance'] < distance]

    # compute competitiveness 
    df_zom_zom['comp_score'] = df_zom_zom.apply(lambda x:compute_competitiveness(x['encode'],x['encode2']),axis=1)

    # remove same ID 
    df_zom_zom = df_zom_zom[df_zom_zom['index'] != df_zom_zom['index2']]

    # aggregate 
    df_zom_zom_agg_raw = df_zom_zom.groupby('index').agg({'index2':'nunique'}).reset_index()
    df_zom_zom_agg_raw.columns = ['index','n_total']

    # filter distance 
    df_zom_zom =  df_zom_zom[df_zom_zom['comp_score'] > 0.4 ]

    # count aggregate 
    df_zom_zom_agg_filter = df_zom_zom.groupby('index').agg({'index2':'nunique','rating2':'mean'}).reset_index()
    df_zom_zom_agg_filter.columns = ['index','n_total_filter'.format(distance),'avg_n_rest']

    # count overall 
    result = pd.merge(df_zom_zom_agg_filter,df_zom_zom_agg_raw, left_on='index', right_on='index')

    # relevant competitiveness / total 
    result['n_compt_{}'.format(distance)] = result['n_total_filter'] / result['n_total']

    result = result[['index','n_compt_{}'.format(distance),'avg_n_rest']]

    return result 
    
def compute_competitiveness(rest_a, rest_b):

    # convert to list 
    list_a = list(map(lambda x:int(x), list(rest_a)))
    list_b = list(map(lambda x:int(x), list(rest_b)))

    # convert to array 
    arr_a = np.array(list_a)
    arr_b = np.array(list_b)

    # compute similarity 
    cosine = np.dot(arr_a,arr_b)/(norm(arr_a)*norm(arr_b))

    return cosine


# df_tmp = pd.read_csv('data/3_clean_zomato_gof_500.csv')
# df_tmp = df_tmp.drop(['url', 'rest_price_idr', 'review', 'lat', 'long', 'is_chain',
#        'rating', 'new_code_res_type', 'rank_res_type', 'new_code_fac',
#        'rank_fac', 'geohash'],axis=1)

# geo_cols = df_tmp.columns[1:]
# df_tmp.head()

# distance=500
# df_tmp['en_{}'.format(distance)] = df_tmp[geo_cols].apply(lambda x:compute_entropy(x),axis=1)

# df_tmp['en_500'].plot(kind='hist')

# load dataset 

df_poi = pd.read_csv('../google_poi.csv')


df_zomato = pd.read_csv('data/3_clean_zomato_feat_ori.csv')
# df_zomato = pd.read_csv('data/3_jbdk_clean_zomato_feat_ori.csv')

df_zomato_type = pd.read_csv('data/2_res_type.csv')

df_zomato = pd.merge(df_zomato, df_zomato_type[['index','encode']], left_on='index', right_on='index', how='left')

# preprocessing POI

df_poi['geohash6'] = df_poi['geohash_poi_8'].apply(lambda x:x[:6])

#filter jakarta 
df_jak_poi = df_poi[(df_poi['city'].str.contains("Jak|Bogor|Depok|Kota_Bekasi|Kab_Tanggerang|Kota_Tanggerang|Kota_Tangsel|Kab_Bekasi"))]
df_jak_poi = df_jak_poi.reset_index()

# preprocessing zomato 
df_zomato['geohash'] = df_zomato.apply(lambda x:geohash.encode(x['lat'],x['long'],6),axis=1)

# perform geo extraction zomato and POI 

# density 
distance = 1000
df_gof_dens = density_gof(df_zomato, df_jak_poi, distance)
print('density done')

# entropy 
geo_cols = df_gof_dens.columns[1:]
df_gof_dens['en_{}'.format(distance)] = df_gof_dens[geo_cols].apply(lambda x:compute_entropy(x),axis=1)
print('entropy done')

# find surrounding competitor
df_gof_compt = compt_gof(df_zomato,df_zomato,distance)
print('competitiveness done')

# merge with other features 
df_gof_dens = pd.merge(df_gof_dens,df_gof_compt,how='left',left_on='d_{}_index'.format(distance),right_on='index')
df_gof_dens = df_gof_dens.drop('index',axis=1)

# merge with main data

df_zomato_gof_dens = pd.merge(df_zomato, df_gof_dens, left_on='index', right_on='d_{}_index'.format(distance), how='left')
df_zomato_gof_dens = df_zomato_gof_dens.fillna(0)
df_zomato_gof_dens = df_zomato_gof_dens.drop('d_{}_index'.format(distance),axis=1)

df_zomato_gof_dens.head()

df_zomato_gof_dens.shape

df_zomato_gof_dens.to_csv('data/3_clean_zomato_gof_ori_{}.csv'.format(distance),index=False)
# df_zomato_gof_dens.to_csv('data/3_clean_zomato_gof_{}.csv'.format(distance),index=False)
# df_zomato_gof_dens.to_csv('data/3_jbdk_clean_zomato_gof_{}.csv'.format(distance),index=False)
