==> 7_stats_test.ipynb <==
import pandas as pd
import numpy as np
from scipy.stats import f_oneway
from statsmodels.stats.multicomp import pairwise_tukeyhsd

# Manual Test 

# MAE 
# zomato = [0.12398,	0.12298,	0.12642]
# zomato_1000 = [0.11915, 0.12145, 0.1255]
# zomato_500 = [0.12006,	0.12244,	0.12491]
# zomato_250 = [0.12001,	0.12278,	0.12174]


# MSE 
# zomato = [0.02546,	0.02467,	0.02615]
# zomato_1000 = [0.02324,	0.02404,	0.0258]
# zomato_500 = [0.02362,	0.02447,	0.02496]
# zomato_250 = [0.02357,	0.02432,	0.02602]

# pearson
zomato = [0.41264,	0.4076,	0.41256]
zomato_1000 = [0.47026,	0.43416,	0.4084]
zomato_500 = [0.45083,	0.41555,	0.4395]
zomato_250 = [0.45421,	0.42217,	0.42244]

f_oneway(zomato, zomato_1000, zomato_500, zomato_250)

score_list = []
score_list.extend(zomato)
print(len(score_list))
score_list.extend(zomato_1000)
print(len(score_list))
score_list.extend(zomato_500)
print(len(score_list))
score_list.extend(zomato_250)
print(len(score_list))

df = pd.DataFrame(
    {'dataset' : ['zomato']*3 + ['zomato_1000']*3 + ['zomato_500']*3 + ['zomato_250']*3  ,
    'score' : score_list
    }
)

df

# perform Tukey's test
tukey = pairwise_tukeyhsd(endog=df['score'],
                          groups=df['dataset'],
                          alpha=0.1)

print(tukey)

# Cross validation Data 

df = pd.read_excel('report/ml_perform_crossval_raw_2022-11-23.xlsx')

df_zomato = df[df['dataset']=='zomato_only']
df_zomato_poi_250 = df[df['dataset']=='zomato_poi_250']
df_zomato_poi_500 = df[df['dataset']=='zomato_poi_500']
df_zomato_poi_1000 = df[df['dataset']=='zomato_poi_1000']

df_zomato_mse = pd.concat([df_zomato[['test_neg_mean_squared_error','dataset']] , df_zomato_poi_250[['test_neg_mean_squared_error','dataset']]
                    , df_zomato_poi_500[['test_neg_mean_squared_error','dataset']] , df_zomato_poi_1000[['test_neg_mean_squared_error','dataset']]
])


df_zomato_mae = pd.concat([df_zomato[['test_neg_mean_absolute_error','dataset']] , df_zomato_poi_250[['test_neg_mean_absolute_error','dataset']]
                    , df_zomato_poi_500[['test_neg_mean_absolute_error','dataset']] , df_zomato_poi_1000[['test_neg_mean_absolute_error','dataset']]
])

df_zomato_pearson = pd.concat([df_zomato[['test_Pearson','dataset']] , df_zomato_poi_250[['test_Pearson','dataset']]
                    , df_zomato_poi_500[['test_Pearson','dataset']] , df_zomato_poi_1000[['test_Pearson','dataset']]
])

# df_zomato_pearson

## MAE

f_oneway(df_zomato['test_neg_mean_absolute_error'], df_zomato_poi_250['test_neg_mean_absolute_error'], df_zomato_poi_500['test_neg_mean_absolute_error'], df_zomato_poi_1000['test_neg_mean_absolute_error'])

# perform Tukey's test
tukey = pairwise_tukeyhsd(endog=df_zomato_mae['test_neg_mean_absolute_error'],
                          groups=df_zomato_mae['dataset'],
                          alpha=0.05)

print(tukey)

## MSE

f_oneway(df_zomato['test_neg_mean_squared_error'], df_zomato_poi_250['test_neg_mean_squared_error'], df_zomato_poi_500['test_neg_mean_squared_error'], df_zomato_poi_1000['test_neg_mean_squared_error'])

# perform Tukey's test
tukey = pairwise_tukeyhsd(endog=df_zomato_mse['test_neg_mean_squared_error'],
                          groups=df_zomato_mse['dataset'],
                          alpha=0.05)

print(tukey)

# Pearson

f_oneway(df_zomato['test_Pearson'], df_zomato_poi_250['test_Pearson'], df_zomato_poi_500['test_Pearson'], df_zomato_poi_1000['test_Pearson'])

# perform Tukey's test
tukey = pairwise_tukeyhsd(endog=df_zomato_pearson['test_Pearson'],
                          groups=df_zomato_pearson['dataset'],
                          alpha=0.05)

print(tukey)
