==> 99_data_explore.ipynb <==
import pandas as pd 
from sklearn.model_selection import train_test_split

# Split train and test dataset 

input_file = 'data/3_clean_zomato_feat_ori_sampling.csv'
train_file = 'data/3_clean_zomato_feat_ori_sampling_train.csv'
test_file  = 'data/3_clean_zomato_feat_ori_sampling_test.csv'

df_input = pd.read_csv(input_file)
target_column = 'rating'
X_train, X_test, y_train, y_test = train_test_split(df_input.drop(target_column, axis=1), df_input[target_column], test_size=0.20, random_state=42)

X_train[target_column] = y_train
X_test[target_column] = y_test


X_train.to_csv(train_file,index=False)
X_test.to_csv(test_file,index=False)

X_train.shape

# combine jakarta and outer jakarta dataset 
df_bod = pd.read_csv('data/2_bodebek_restaurant_detail3.csv')
df_jak = pd.read_csv('data/2_jak_rest_detail.csv')

df_combine = pd.concat([df_jak,df_bod]).reset_index()

df_combine.to_csv('data/2_jbdtk_rest_detail.csv',index=False)

df_combine.head()

# compare raw data 

df1 = pd.read_pickle('../../dataset/restaurant_raw.pickle')
df2 = pd.read_pickle('../../dataset/restaurant_raw2.pickle')
df3 = pd.read_pickle('../../dataset/restaurant_raw3.pickle')

df1_exp = df1.explode('link')
df2_exp = df2.explode('link')
df3_exp = df3.explode('link')

df1_exp.shape, df2_exp.shape, df3_exp.shape

df_union = pd.concat([df1_exp,df2_exp,df3_exp]).drop_duplicates()

df_union.shape

df_union.columns = ['district','rest_url','count']

df_union.head()

df_union.to_csv('jakarta_rest.csv',index=False)

tes_df = pd.read_csv('jakarta_restaurant_detail.csv')
tes_df2 = pd.read_csv('jakarta_restaurant_detail2.csv')
tes_df3 = pd.read_csv('jakarta_restaurant_detail3.csv')
tes_df_uni = pd.concat([tes_df,tes_df2,tes_df3]).drop_duplicates()

tes_df_uni.isnull().sum()

tes_df.shape, tes_df2.shape, tes_df3.shape, tes_df_uni.shape

tes_df.head()

tes_df2.head()

tes_df_uni.to_csv('3_jak_rest_detail.csv',index=False)
