==> 99_POI_explore.ipynb <==
import pandas as pd 

df = pd.read_csv('../google_poi.csv')

df.head()

df['city'].value_counts()

# df[df['city'].str.contains("Jak")]['city'].value_counts()

df_jak = df[df['city'].str.contains("Jak")]

df_jak['poi_type'].value_counts().shape

df_jak_res = df[(df['city'].str.contains("Jak")) & (df['poi_type'].str.contains("cafe|restaurant|meal_takeaway"))]
df_jak_poi = df[(df['city'].str.contains("Jak"))]

df_jak_res.head()

df_jak_res['poi_type'].value_counts()

df_jak_poi.shape

df_jak_poi['poi_type'].value_counts()
